+++ predeploy.ps1
@@ -20,3 +20,11 @@ $contents = $contents + @"
 }
 "@
 $contents | Out-File "./approot/src/Kraken/appsettings.deploy.json"
+
+If (-Not [string]::IsNullOrEmpty($StdoutLogEnabled)) {
+	$config = [xml](Get-Content "./wwwroot/web.config")
+	$node = $config.SelectNodes("configuration/system.webServer/httpPlatform")
+	$node.SetAttribute("stdoutLogEnabled", $StdoutLogEnabled)
+	$config.Save("./wwwroot/web.config")
+}
+New-Item -ItemType Directory -Path "./logs"