+++ src/Kraken/Services/OctopusAuthenticationProxy.cs
@@ -28,7 +28,7 @@
             {
                 _repository.Users.SignIn(loginCommand);
             }
-            catch
+            catch (OctopusValidationException)
             {
                 user = null;
                 return false;