+++ src/Kraken/wwwroot/js/services/releaseBatches.js
@@ -108,14 +108,18 @@
                 return data;
             });
         },
-        getProgression : function(idOrName) {
+        getProgression: function (idOrName) {
             return $.ajax({
                 type: 'GET',
                 url: context.basePath + 'api/releasebatches/' + encodeURI(idOrName) + '/getprogression',
+                traditional: true,
                 contentType: 'application/json'
             }).then(function (data) {
                 return data;
             });
+        },
+        getBatchEnvironments: function (idOrName, permission) {
+            return $.get(context.basePath + 'api/releasebatches/' + encodeURI(idOrName) + '/getbatchenvironments');
         }
     };