+++ RELEASENOTES.md
@@ -1,3 +1,3 @@
 ### Bug Fixes
 
-* appveyor artifact push fix ([53111a8](https://github.com/zywave/OctopusDeploy-Kraken/commit/53111a8)), closes [#31](https://github.com/zywave/OctopusDeploy-Kraken/issues/31)
+* logs file powershell fix ([fd3fccf](https://github.com/zywave/OctopusDeploy-Kraken/commit/fd3fccf))